backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
minio==7.1.15
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
asyncpg==0.29.0
pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
aiofiles==23.2.1
